install phase fixed up

+++ mk_live_usb.sh
@@ -2,6 +2,8 @@
 . ./config_params.sh
 
 ERASE_DEV=0
+FORMAT_DEV=0
+INSTALL=0
 DEV=/dev/sdd
 LIVE_PART=${DEV}1
 WORK_PART=${DEV}2
@@ -17,11 +19,15 @@ echo -e "
 \e[1mOPTIONS\e[0m
 
   -e, --erase
-     erase the device
+     erase the device and create partitions
+  -f, --format
+     format the partitions
   -d, --dev
      device path
   -i, --image
      live image iso
+  -I, --install
+     install the live system
 "
 }
 
@@ -49,11 +55,20 @@ Are you sure you want to continue? [N/y] "
 }
 
 cleanup() {
+  printf "\e[92mCleaning up...\e[0m\n"
   [ -d "${USB_KEY}" ] && umount ${USB_KEY} && rmdir ${USB_KEY}
   [ -d "${LIVE_DIR}" ] && umount ${LIVE_DIR} && rmdir ${LIVE_DIR}
+  umount ${LIVE_PART}
+  umount ${WORK_PART}
 }
 
 erase_dev() {
+  if [ $(mount | grep ${DEV} | wc -l) != "0" ]; then
+    printf "\n\e[91mERROR: \e[0mFailed to erase ${DEV}\n"
+    printf "\e[33mHINT: \e[0mtry to umount the device\n"
+    exit 1
+  fi
+
   printf "\e[92mErasing ${DEV}\e[0m --->  "
   dd if=/dev/zero of=${DEV} bs=1024k > /dev/null 2>&1 &
   spin $!
@@ -64,7 +79,7 @@ erase_dev() {
   parted -s ${DEV} "mkpart primary fat32 0 2G" > /dev/null 2>&1 &
   spin $!
   parted -s ${DEV} "set 1 boot on" > /dev/null 2>&1 &
-  printf "\e[92mCreating workspace (ext2) partition on ${WORK_PART}\e[0m --->  "
+  printf "\e[92mCreating workspace (ext3) partition on ${WORK_PART}\e[0m --->  "
   parted -s ${DEV} "mkpart primary ext3 2G -1" > /dev/null 2>&1 &
   spin $!
   printf "\e[92mInstalling MBR on ${DEV}\e[0m --->  "
@@ -75,15 +90,25 @@ erase_dev() {
 format() {
   printf "\e[92mFormatting ${LIVE_PART}\e[0m --->  "
   mkfs.msdos -F32 ${LIVE_PART} > /dev/null 2>&1 &
+  if [ $(mount | grep ${LIVE_PART} | wc -l) != "0" ]; then
+    printf "\n\e[91mERROR: \e[0mFailed to create FAT32 filesystem on ${LIVE_PART}\n"
+    printf "\e[33mHINT: \e[0mtry to umount the device\n"
+    exit 1
+  fi
   spin $!
   printf "\e[92mFormatting ${WORK_PART}\e[0m --->  "
   mkfs.ext3 ${WORK_PART} > /dev/null 2>&1 &
+  if [ $(mount | grep ${WORK_PART} | wc -l) != "0" ]; then
+    printf "\n\e[91mERROR: \e[0mFailed to create ext3 filesystem on ${WORK_PART}\n"
+    printf "\e[33mHINT: \e[0mtry to umount the device\n"
+    exit 1
+  fi
   spin $!
 }
 
 install_image() {
   printf "\e[92mInstalling bootloader\e[0m --->  "
-  syslinux ${LIVE_PART} > /dev/null 2>&1 &
+  syslinux -i ${LIVE_PART} > /dev/null 2>&1 &
   spin $!
 
   mkdir ${USB_KEY}
@@ -95,14 +120,16 @@ install_image() {
   cd ${LIVE_DIR}
   cp -R * ${USB_KEY} > /dev/null 2>&1 &
   spin $!
+  cd ${WORK_DIR}
 
   printf "\e[92mConfiguring isolinux\e[0m --->  "
-  cp isolinux/* ${USB_KEY} > /dev/null 2>&1 &
+  #cp isolinux/* ${USB_KEY} > /dev/null 2>&1 &
+  mv ${USB_KEY}/isolinux ${USB_KEY}/syslinux > /dev/null 2>&1 &
   spin $!
-  cd ${USB_KEY}
-  mv isolinux.cfg syslinux.cfg
+  mv ${USB_KEY}/syslinux/isolinux.cfg ${USB_KEY}/syslinux/syslinux.cfg
+}
 
-  cd ${WORK_DIR}
+configure_workspace() {
 }
 
 if [ $(id -u) != 0 ]; then 
@@ -122,6 +149,12 @@ while [ $# -gt 0 ]; do
   -e | --erase)
     ERASE_DEV=1
     ;;
+  -f | --format)
+    FORMAT_DEV=1
+    ;;
+  -I | --install)
+    INSTALL=1
+    ;;
   -d | --dev)
     DEV=$1
     shift
@@ -139,12 +172,12 @@ done
 
 askCONF
 
-if [ $ERASE_DEV -eq 1 ]; then
-  erase_dev
-fi
+trap cleanup SIGINT SIGTERM
 
-format
-install_image
+[ $ERASE_DEV -eq 1 ] && erase_dev
+[ $FORMAT_DEV -eq 1 ] && format
+[ $INSTALL -eq 1 ] && install_image
+
 cleanup
 
-printf "\e[92mLive USB is ready!\e[0\nm"
+printf "\e[92mUSB drive is ready!\e[0\nm"
